Extract out the module dependency stuff to the module file only (leaving hooks for others in case of future desires)
