+++ config/groovy/facade.groovy
@@ -7,4 +7,8 @@ class facade {
 
     def excludedItems = ["PC", "TeraEd"]
 
+    // Facades currently do not care about dependencies
+    String[] findDependencies(File targetDir) {
+        return []
+    }
 }
